refactor: goldlog
