+++ packages/iceworks-server/src/app/controller/goldlog.ts
@@ -1,33 +1,8 @@
 import { controller, provide, post } from 'midway-mirror';
-import * as request from 'request-promise-native';
 import storage from '../../lib/storage';
+import goldlog from '../../lib/goldlog';
 
 const { checkAliInternal } = require('ice-npm-utils');
-
-async function record(data) {
-  const dataKeyArray = Object.keys(data);
-  const gokey = dataKeyArray.reduce((finnalStr, currentKey, index) => {
-    const currentData =
-        typeof data[currentKey] === 'string'
-          ? data[currentKey]
-          : JSON.stringify(data[currentKey]);
-    return `${finnalStr}${currentKey}=${currentData}${
-      dataKeyArray.length - 1 === index ? '' : '&'
-    }`;
-  }, '');
-
-  await request({
-    method: 'post',
-    url: 'http://gm.mmstat.com/iceteam.iceworks.log3',
-    json: true,
-    body: {
-      cache: Math.random(),
-      gmkey: 'CLK',
-      gokey: encodeURIComponent(gokey),
-      logtype: '2',
-    },
-  });
-}
 
 @provide()
 @controller('/api/goldlog')
@@ -37,13 +12,7 @@ export class GoldlogController {
   public async record(ctx) {
     if (ctx.request.body) {
       const data = ctx.request.body;
-
-      try {
-        await record(data);
-      } catch (error) {
-        error.name = 'goldlog-error';
-        ctx.logger.error(error);
-      }
+      await goldlog(data);
     }
 
     ctx.body = {
@@ -57,14 +26,18 @@ export class GoldlogController {
     const nowtDate = new Date().toDateString();
 
     const lastDate = storage.get('lastDate');
+    const locale = storage.get('locale');
+    const theme = storage.get('theme');
     if(nowtDate !== lastDate) {
       storage.set('lastDate', nowtDate);
-      record({
+      await goldlog({
         namespace: 'home',
         module: 'log',
         action: 'dau',
         data: {
           group: isAlibaba ? 'alibaba' : 'outer',
+          locale,
+          theme,
         },
       });
     }
